appointment logic and database interaction begin

+++ ScheduleIt/Main.cs
@@ -7,6 +7,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Data;
+using System.Data.Entity;
 using MySql.Data;
 using MySql.Data.MySqlClient;
 using ApptsDb;
@@ -20,22 +21,29 @@ namespace ScheduleIt
             InitializeComponent();
         }
 
+        private int ActiveTab { get; set; }
+
         //buttons start
         private void addButton_Click(object sender, EventArgs e)
         {
-            switch (tabControl1.SelectedIndex)
+            switch (ActiveTab)
             {
                 //Appointments
                 case 0:
-                    AppointmentEdit apptWin = new AppointmentEdit(true);
-                    apptWin.Show();
+                    AppointmentEdit apptEdit = new AppointmentEdit();
+                    apptEdit.ShowDialog();
+                    DialogResult apptEditRes = apptEdit.DialogResult;
+                    if( apptEditRes == DialogResult.OK)
+                    {
+                        _UpdateAppointmentsTbl();
+                    }
                     break;
                 //Customers
                 case 1:
                     CustomerEdit custEditForm = new CustomerEdit();
                     custEditForm.ShowDialog();
-                    DialogResult res = custEditForm.DialogResult;
-                    if( res == DialogResult.OK)
+                    DialogResult custEditRes = custEditForm.DialogResult;
+                    if( custEditRes == DialogResult.OK)
                     {
                         _UpdateCustomersTbl();
                     }
@@ -45,19 +53,27 @@ namespace ScheduleIt
 
         private void deleteButton_Click(object sender, EventArgs e)
         {
-            switch (tabControl1.SelectedIndex)
+            switch (ActiveTab)
             {
                 //Appointments
                 case 0:
-                    AppointmentEdit apptWin = new AppointmentEdit(true);
-                    apptWin.Show();
+                    var apptId = _GetSelectedId();
+                    DialogResult deleteApptConfirm = MessageBox.Show("Are you sure you want to delete?", "Delete Appointment", MessageBoxButtons.YesNo);
+                    if(deleteApptConfirm == DialogResult.Yes)
+                    {
+                        DataAccess.DeleteAppointment(apptId);
+                    }
+                    _UpdateAppointmentsTbl();
                     break;
                 //Customers
                 case 1:
-                    int rowIndex = customersDataGrid.SelectedRows[0].Index;
-                    DataGridViewRow selectedRow = customersDataGrid.Rows[rowIndex];
-                    var custId = int.Parse(selectedRow.Cells["Id"].Value.ToString());
-                    DataAccess.DeleteCustomer(custId);
+                    int custId = _GetSelectedId();
+                    DialogResult deleteCustConfirm = MessageBox.Show("Are you sure you want to delete?", "Delete Customer", MessageBoxButtons.YesNo);
+                    if(deleteCustConfirm == DialogResult.Yes)
+                    {
+                        DataAccess.DeleteCustomer(custId);
+                    }
+
                     _UpdateCustomersTbl();
                     break;
             }
@@ -65,18 +81,22 @@ namespace ScheduleIt
 
         private void modifyButton_Click(object sender, EventArgs e)
         {
-            switch (tabControl1.SelectedIndex)
+            switch (ActiveTab)
             {
                 //Appointments
                 case 0:
-                    AppointmentEdit apptWin = new AppointmentEdit(true);
-                    apptWin.Show();
+                    var apptId = _GetSelectedId();
+                    AppointmentEdit apptWin = new AppointmentEdit(apptId);
+                    apptWin.ShowDialog();
+                    DialogResult apptRes = apptWin.DialogResult;
+                    if (apptRes == DialogResult.OK)
+                    {
+                        _UpdateCustomersTbl();
+                    }
                     break;
                 //Customers
                 case 1:
-                    int rowIndex = customersDataGrid.SelectedRows[0].Index;
-                    DataGridViewRow selectedRow = customersDataGrid.Rows[rowIndex];
-                    var custId = int.Parse(selectedRow.Cells["Id"].Value.ToString());
+                    var custId = _GetSelectedId();
                     CustomerEdit custEditForm = new CustomerEdit(custId);
                     custEditForm.ShowDialog();
                     DialogResult res = custEditForm.DialogResult;
@@ -92,8 +112,9 @@ namespace ScheduleIt
         private void tabControl1_SelectedIndexChanged(object sender, EventArgs e)
         {
             TabControl tab = sender as TabControl;
+            ActiveTab = tab.SelectedIndex;
             
-            if (tab.SelectedIndex == 1)
+            if (ActiveTab == 1)
             {
                 try
                 {
@@ -116,6 +137,27 @@ namespace ScheduleIt
             List<CustomerDisplay> custs = DataAccess.GetCustomerDisplay();
             customersDataGrid.DataSource = custs;
             customersDataGrid.Columns["Id"].Visible = false;
+        }
+
+        private void _UpdateAppointmentsTbl()
+        {
+            List<AppointmentDisplay> appts = DataAccess.GetAppointmentDisplay(GlobalVar.LoggedInUser.UserId);
+            appointmentGridView.DataSource = appts;
+            appointmentGridView.Columns["Id"].Visible = false;
+        }
+
+        private void Main_Load(object sender, EventArgs e)
+        {
+            _UpdateAppointmentsTbl();
+        }
+
+        private int _GetSelectedId()
+        {
+            var grid = ActiveTab == 1 ? customersDataGrid : appointmentGridView;
+            int rowIndex = grid.SelectedRows[0].Index;
+            DataGridViewRow selectedRow = grid.Rows[rowIndex];
+
+            return int.Parse(selectedRow.Cells["Id"].Value.ToString());
         }
     }
 }
